Move socialite functions into their own files; Updated tests and some use cases
