+++ src/index.ts
@@ -1,6 +1,8 @@
-import * as socialite from './socialite';
+import buildUrl from './buildUrl';
+import extractUser from './extractUser';
+import isValidDomain from './isValidDomain';
 
 export * from './types/socials';
-export * from './socialite';
-
+export { buildUrl, extractUser, isValidDomain };
+const socialite = { buildUrl, extractUser, isValidDomain };
 export default socialite;
